requirements.txt
alembic==0.9.3
async-generator==1.10
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
click==6.7
cryptography==36.0.1
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
h11==0.12.0
idna==3.3
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.1.1
outcome==1.1.0
Pillow==9.0.0
pkg-resources==0.0.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.6.1
python-editor==1.0.3
RPi.GPIO==0.7.0
selenium==3.14.0
six==1.10.0
sniffio==1.2.0
sortedcontainers==2.4.0
spidev==3.5
SQLAlchemy==1.1.11
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
visitor==0.1.3
Werkzeug==0.12.2
wsproto==1.0.0
WTForms==2.1
